Fixed profile.

+++ Controllers/ProfileController.cs
@@ -8,7 +8,7 @@ using MySql.Data.MySqlClient;
 namespace EffectiveWebProg.Controllers
 {
 
-    public class ProfileController : Controller
+    public class ProfileController : BaseController
     {
         private readonly ApplicationDbContext _db;
         private readonly string connectionString = "server=mysql-webprogramming1-sit-cc31.c.aivencloud.com;port=19112;database=Munch;uid=avnadmin;pwd=AVNS_HsKVnqOod_xgB4OJwUT;sslmode=Required";
@@ -140,7 +140,7 @@ namespace EffectiveWebProg.Controllers
             return View(user);
         }
 
-        [HttpPost("EditProfile")]
+        [HttpPost("Profile/EditProfile")]
         public async Task<IActionResult> EditProfilePost(IFormFile UserProfilePic, string UserName, string UserUsername, string UserEmail, int UserContactNum)
         {
             var sessionID = Guid.Parse(HttpContext.Session.GetString("SSID") ?? "");
@@ -158,20 +158,24 @@ namespace EffectiveWebProg.Controllers
                 return RedirectToAction("EditProfile");
             }
 
-            var uploadsFolder = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images", "UserProfilePics");
-            string filePath;
-            if (user.UserProfilePic != null)
-            {
-                filePath = Path.Combine(uploadsFolder, user.UserProfilePic);
-                System.IO.File.Delete(filePath);
+            if (UserProfilePic != null)
+            {    
+                var uploadsFolder = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "Images", "UserProfilePics");
+                string filePath;
+                if (user.UserProfilePic != null)
+                {
+                    filePath = Path.Combine(uploadsFolder, user.UserProfilePic);
+                    System.IO.File.Delete(filePath);
+                }
+
+                var filename = $"{user.UserID}{Path.GetExtension(UserProfilePic.FileName)}";
+                filePath = Path.Combine(uploadsFolder, filename);
+                using var fileStream = new FileStream(filePath, FileMode.Create);
+                await UserProfilePic.CopyToAsync(fileStream);
+
+                user.UserProfilePic = filename;
             }
 
-            var filename = $"{user.UserID}{Path.GetExtension(UserProfilePic.FileName)}";
-            filePath = Path.Combine(uploadsFolder, filename);
-            using var fileStream = new FileStream(filePath, FileMode.Create);
-            await UserProfilePic.CopyToAsync(fileStream);
-
-            user.UserProfilePic = filename;
             user.UserName = UserName;
             user.UserUsername = UserUsername;
             user.UserEmail = UserEmail;
